Moved workspaces-items
