+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/workspaces-items/presentation.tsx
@@ -14,9 +14,9 @@
  **/
 import * as React from 'react'
 import styled from 'styled-components'
-import { CustomElement } from '../../styles/mixins'
-import WorkspaceItemContainer from '../../workspace-item'
-import { Button, buttonTypeEnum } from '../button'
+import { CustomElement } from '../styles/mixins'
+import WorkspaceItemContainer from '../workspace-item'
+import { Button, buttonTypeEnum } from '../presentation/button'
 import { hot } from 'react-hot-loader'
 
 type Props = {
